+++ packages/@runlightyear/cli/src/shared/createAuthorizerActivity.ts
@@ -1,0 +1,55 @@
+import fetch from "node-fetch";
+import { getApiKey, getBaseUrl, getEnvName } from "@runlightyear/lightyear";
+
+export interface CreateAuthorizerActivityProps {
+  customAppName: string;
+  type:
+    | "GET_AUTH_REQUEST_URL"
+    | "REQUEST_ACCESS_TOKEN"
+    | "REFRESH_ACCESS_TOKEN";
+  status: "SUCCEEDED" | "FAILED";
+  logs?: any;
+}
+
+export default async function createAuthorizerActivity(
+  props: CreateAuthorizerActivityProps
+) {
+  const { customAppName, type, status, logs } = props;
+
+  const baseUrl = getBaseUrl();
+  const envName = getEnvName();
+  const apiKey = getApiKey();
+
+  const activityResponse = await fetch(
+    `${baseUrl}/api/v1/envs/${envName}/custom-apps/${customAppName}/authorizer/activities`,
+    {
+      method: "POST",
+      headers: {
+        Authorization: `apiKey ${apiKey}`,
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        type,
+        status,
+        logs,
+      }),
+    }
+  );
+
+  if (activityResponse.ok) {
+    console.debug("Uploaded authorizer activity result");
+    const json = await activityResponse.json();
+    if (!json.id) {
+      throw new Error("Missing Authorizer activity id");
+    }
+    return json.id as string;
+  } else {
+    console.error(
+      "Failed to create authorizer activity: ",
+      activityResponse.status,
+      activityResponse.statusText
+    );
+    console.error(JSON.stringify(await activityResponse.json(), null, 2));
+    throw new Error("Failed to create authorizer activity");
+  }
+}