+++ packages/@runlightyear/gmail/src/Gmail.ts
@@ -1,15 +1,19 @@
-import { AuthConnectorProps, RestConnector } from "@runlightyear/lightyear";
+import { RestConnector, RestConnectorProps } from "@runlightyear/lightyear";
 
 /**
  * @alpha
  */
-export interface GmailProps extends AuthConnectorProps {}
+export interface GmailProps extends RestConnectorProps {}
 
 /**
  * @alpha
  */
 export class Gmail extends RestConnector {
   constructor(props: GmailProps) {
-    super({ ...props, baseUrl: "https://gmail.googleapis.com" });
+    super(props);
+  }
+
+  getBaseUrl() {
+    return "https://gmail.googleapis.com";
   }
 }