+++ packages/@runlightyear/lightyear/src/handler/handleGetAuthRequestUrl.ts
@@ -1,7 +1,5 @@
 import { handlerResult } from "./handlerResult";
-import { unsubscribe } from "../subscriptionActivities";
-import { getWebhookData } from "../base/webhook";
-import { oAuthIndex } from "../base/oauth";
+import { authorizerIndex } from "../base/authorizer";
 import { OAuthConnector } from "../connectors/OAuthConnector";
 import { getAuthData, updateAuthDataState } from "../base/auth";
 import invariant from "tiny-invariant";
@@ -15,11 +13,13 @@ export async function handleGetAuthRequestUrl({
   customApp,
   authName,
 }: HandleGetAuthRequestUrl) {
+  console.debug("Getting auth request url");
+
   if (!customApp) {
     return handlerResult(400, "Missing customApp");
   }
 
-  const item = oAuthIndex[customApp];
+  const item = authorizerIndex[customApp];
 
   await updateAuthDataState({ customAppName: customApp, authName });