+++ packages/@runlightyear/lightyear/src/handler/handleRequestAccessToken.ts
@@ -1,5 +1,5 @@
 import { handlerResult } from "./handlerResult";
-import { oAuthIndex } from "../base/oauth";
+import { authorizerIndex } from "../base/authorizer";
 import { OAuthConnector } from "../connectors/OAuthConnector";
 import { getAuthData, updateAuthDataState } from "../base/auth";
 import invariant from "tiny-invariant";
@@ -15,6 +15,8 @@ export async function handleRequestAccessToken({
   authName,
   code,
 }: HandleRequestAccessTokenProps) {
+  console.debug("Requesting access token");
+
   if (!customAppName) {
     return handlerResult(400, "Missing customApp");
   }
@@ -27,7 +29,7 @@ export async function handleRequestAccessToken({
     return handlerResult(400, "Missing code");
   }
 
-  const item = oAuthIndex[customAppName];
+  const item = authorizerIndex[customAppName];
 
   const authData = await getAuthData({
     customAppName: customAppName,
@@ -46,6 +48,8 @@ export async function handleRequestAccessToken({
   });
 
   await connector.requestAccessToken(code);
+
+  console.info("Requested access token successfully");
 
   return handlerResult(200, "Success");
 }