+++ packages/@runlightyear/salesforce/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @runlightyear/salesforce
 
+## 0.4.0
+
+### Minor Changes
+
+- 178a7d78b: Allow specification of scopes
+- 178a7d78b: Implement getBaseUrl
+
+### Patch Changes
+
+- Updated dependencies [2914452a6]
+- Updated dependencies [2914452a6]
+  - @runlightyear/lightyear@0.37.0
+
 ## 0.3.7
 
 ### Patch Changes